requirements.txt
unstructured[all-docs]
pillow
lxml
tiktoken
langchain
langchain-community
langchain-openai
langchain-groq
python-dotenv
faiss-cpu
pysqlite3-binary
tqdm
notebook
ipywidgets
chromadb==0.4.24
numpy<2.0
streamlit
